build: Make most entities private/hidden.
- Moves the CLI's main() to the library to avoid exposing
  a bunch of types and functions as private.
- Marks the must-be-exported API for internal crates with
  #[doc(hidden)] to avoid usage by someone else.

We want to re-work the API for the future, so this is a first
step before we can do renaming internally.
